requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.0,<4.0
PyMySQL>=1.1.0,<2.0
python-dotenv>=1.0,<2.0